Revamped the Menu system
Changed stack so that it holds a pair, in which the first element
is a pointer to the current PARENT gui, instead of the FOCUSED gui like
before. The second element is the index of FOCUSED gui, in the PARENT's
CHILDREN vector.

I was having problems before when I was storing a pointer to FOCUSED
gui in the gui_stack. I think this is much simpler and generic.